Move lite module under indicators
